requirements.txt
# requirements.txt
langchain==0.1.20
langchain-community==0.0.29
streamlit==1.32.0
fastapi==0.109.0
uvicorn==0.27.0
pinecone
python-multipart==0.0.9
openai==1.12.0
sentence-transformers==2.6.1
pypdf==4.0.1
python-pptx==0.6.23
openpyxl==3.1.2
transformers==4.38.2
torch==2.2.1
groq==0.4.1
pydantic==2.6.4